+++ ChangeLog.md
@@ -4,6 +4,7 @@
 ## [Unrelease]
 ### Added
 - Implement method `myDataset::load()` to load mnist dataset.
+- Add method into the class `myDataset` to get the fields.
 
 ### Changed